feat(group-clear-restore) add group clear and restore functions to fabric utils
